Add Controller for Comments and implment new mehtod in Comment Repository

+++ JavaServer/src/main/java/com/project/piss/models/Comment.java
@@ -18,16 +18,16 @@ public class Comment {
     private String content;
 
     @Column(name = "sight_id")
-    private int sightId;
+    private Long sightId;
 
     @Column(name = "author_id")
-    private int authorId;
+    private Long authorId;
 
     public Comment() {
 
     }
 
-    public Comment(String content, int sightId, int authorId) {
+    public Comment(String content, Long sightId, Long authorId) {
         this.content = content;
         this.sightId = sightId;
         this.authorId = authorId;
